HSEARCH-1093 Use Java Members directly instead of Hibernate ORM's Getters and PropertyAccessStrategies to access property values
There are plenty of reasons:

* Using gorg.hibernate.property.access.spi.Getter is really complex and
requires hacks to access ORM internals when you want to retrieve the
getter corresponding to a particular property of a particular type.
* Getters do not get created for MappedSuperclasses, forcing us to fall
back to some default PropertyAccessStrategy
* PropertyAccessStrategies do not work for @Transient-annotated properties
(an exception is thrown and they are ignored)
* ORM and we rely on XProperty etc. to retrieve annotations, which makes it
impossible for ORM users to annotate methods with no get/is prefix.
As a result, there is little benefit to using ORM's "property accessor
strategies": it would only allow users to annotate their fields, and
then force ORM and us to access the fields through getters with names
that do not follow the JavaBean conventions. It would not allow them to
annotate these "non-standard" getters directly.