Add ContainerValueExtractorPath to represent a chain of extractors
The main goal is to have a way to represent either "default extractors",
"no extractors" or "the extractors for these classes" in a single
object, which was not previously possible because there was no way to
represent the "default extractors".

This will help to model the associations when we will try to reverse
them to handle automatic reindexing.