Clarify that PojoCasters can only be obtained for raw types
When we have to cast to a generic type, we should be warned exactly
where we do the cast. The way it previously worked, the warning occurred
and was suppressed inside the generic type model, and we didn't even
know we were performing an unchecked cast on the client side.